Refactor to fix code generation and optimize ZIP downloads
The ZIP files were downloaded to the Server's Root, which is not an ideal place for temporary downloads. Instead of the Server Root, the WordPress root (`ABSPATH`) is used instead, to avoid possible conflicts.

Due to the recent renaming of the `theme-pattern-library` repository to the new `theme-components`, the generator code was no longer working.

It was necessary to update the references of the old repository and replace them with the new one in order for the generator to work properly.