Add gulpfile, build process, & docs (fix #63)
This introduces a standard way to build the project's SCSS files and docs
to go along with it. It doesn't require global gulp, so you can run everything
using `npm run build` and `npm start` without needing to `npm install -g gulp`.

This could relate to #42 too, I'm not sure what that issue meant, but
this adds a bit of info on getting the build tools installed and running.