Fix issue with generator failing to build type.
When a type is declared, including stylesheet files inside `assets/stylesheets`,
but not including a `style.scss` file, there was a situation in which the
generator failed to build such type, because the `style.scss` file was missing.

This fix copies and creates the files in place as needed, to make sure the
generator is able to build the type successfully.