Refactor to pass property values to generator filters.
The values of the `bypass_cache` and `logging` properties are being filtered. But, the values passed to these filters are booleans, instead of specifying the original values of the properties.

This commit fixes this and passes the original values to the filters, to keep things consistent and avoid unexpected behavior down the road.