Specify component file in order to replace author in footer correctly.
Themes are currently being generated with "Automattic" in the footer
regardless of user input. This seems to be because the wrong filename is
specified in the generator: it's looking to replace the text in footer.php,
but that text has since been moved to components/footer/site-info.php.

I've updated the generator to look for site-info.php instead. Fixes #150.